Light>01-Color Finished

+++ Chapter2-Lighting/01-Color/src/Camera.h
@@ -70,8 +70,8 @@ public:
 		deltaX *= MouseSensitivity;
 		deltaY *= MouseSensitivity;
 
-		Yaw += deltaX;
-		Pitch -= deltaY;
+		Yaw += (float)deltaX;
+		Pitch -= (float)deltaY;
 
 		if (Pitch > 89.0f)
 			Pitch = 89.0f;
